# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Revolution service API",
        "title": "Revolution service API",
        "termsOfService": "http://github.com",
        "contact": {
            "name": "API Support",
            "url": "Fuck!",
            "email": "U!"
        },
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/clean-all-user": {
            "post": {
                "description": "清除测试账户",
                "tags": [
                    "Test Base"
                ],
                "summary": "清除测试账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/file/download": {
            "post": {
                "description": "通过cid下载文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NFT"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "description": "cid",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.DownloadParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.Context"
                        }
                    }
                }
            }
        },
        "/file/up-to-ipfs": {
            "post": {
                "description": "上传文件到ipfs 返回cid",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NFT"
                ],
                "summary": "上传文件到ipfs",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/generate": {
            "post": {
                "description": "生成测试账户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Base"
                ],
                "summary": "生成测试账户",
                "parameters": [
                    {
                        "description": "生成地址数量",
                        "name": "quantity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.GenerateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/nft/add": {
            "post": {
                "description": "存入nft信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NFT"
                ],
                "summary": "存入nft信息",
                "parameters": [
                    {
                        "description": "add nft",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.AddNftParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/nft/search": {
            "post": {
                "description": "获取nft信息  typa\t1 查询所有的nft   2 所有在售的nft   3所有在游戏中的  4根据key查询自己的nft",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NFT"
                ],
                "summary": "获取nft信息",
                "parameters": [
                    {
                        "description": "nft id",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.SearchParm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.Context": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "errorMsgs"
                },
                "keys": {
                    "type": "object",
                    "additionalProperties": true
                },
                "params": {
                    "type": "Params"
                },
                "request": {
                    "type": "string"
                },
                "writer": {
                    "type": "ResponseWriter"
                }
            }
        },
        "web.AddNftParam": {
            "type": "object",
            "required": [
                "cid",
                "owner",
                "token_id"
            ],
            "properties": {
                "cid": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                }
            }
        },
        "web.DownloadParam": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                }
            }
        },
        "web.GenerateParam": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "web.SearchParm": {
            "type": "object",
            "required": [
                "page",
                "page_size",
                "typa"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "typa": {
                    "type": "integer"
                }
            }
        }
    }
}